+++ test/examples/SampleCrowdsale.test.js
@@ -1,5 +1,4 @@
 const { ether } = require('../helpers/ether');
-const { advanceBlock } = require('../helpers/advanceToBlock');
 const shouldFail = require('../helpers/shouldFail');
 const time = require('../helpers/time');
 const { balanceDifference } = require('../helpers/balanceDifference');
@@ -20,7 +19,7 @@ contract('SampleCrowdsale', function ([_, deployer, owner, wallet, investor]) {
 
   before(async function () {
     // Advance to the next block to correctly read time in the solidity "now" function interpreted by ganache
-    await advanceBlock();
+    await time.advanceBlock();
   });
 
   beforeEach(async function () {
